[mac] Extend frame length from u8 to u16.
This adds support for MAC frames greater than 255 bytes in preparation
for upcoming link layers which can support such long frames.

+++ include/openthread/platform/radio.h
@@ -176,7 +176,7 @@ typedef struct otRadioIeInfo
 typedef struct otRadioFrame
 {
     uint8_t *      mPsdu;    ///< The PSDU.
-    uint8_t        mLength;  ///< Length of the PSDU.
+    uint16_t       mLength;  ///< Length of the PSDU.
     uint8_t        mChannel; ///< Channel used to transmit/receive the frame.
     otRadioIeInfo *mIeInfo;  ///< The pointer to the Header IE(s) related information.
